% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataCOVID.R
\name{getDataCOVID}
\alias{getDataCOVID}
\title{Get COVID-19 Data}
\usage{
getDataCOVID(start = NULL, finish = NULL, country = NULL)
}
\arguments{
\item{start}{enter a start date in mm/dd/yy format. Start date can not be earlier than 01/22/20. Start date can not be later than finish date. If start date is null then start date will be 01/22/20.}

\item{finish}{enter a finish date in mm/dd/yy format. Finish date can not be earlier than start date. If finish date is null then finish date will be the latest date at John-Hopkins CSSE system.}

\item{country}{name of the desired country.}
}
\description{
The function collects the updated  COVID-19 data from the
John Hopkins University
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
