% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaFun.R
\name{kappaFun}
\alias{kappaFun}
\title{Anonymous function approximating the death rate}
\usage{
kappaFun(a, t)
}
\arguments{
\item{a}{parameter vector}

\item{t}{time vector}
}
\value{
No return value, called for side effects
}
\description{
Anonymous function approximating the death rate
}
