% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaFun.R
\name{lambdaFun}
\alias{lambdaFun}
\title{Anonymous function approximating the recovery rate}
\usage{
lambdaFun(a, t)
}
\arguments{
\item{a}{parameter vector}

\item{t}{time vector}
}
\value{
No return value, called for side effects
}
\description{
Anonymous function approximating the recovery rate
}
