% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFun.R
\name{modelFun}
\alias{modelFun}
\title{Model function}
\usage{
modelFun(Y, A, K)
}
\arguments{
\item{Y}{time vector}

\item{A}{the matrix A that is found in: dY/dt = A*Y + F}

\item{K}{the zero matrix for the seven states}
}
\value{
No return value, called for side effects
}
\description{
Model function
}
