% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSEIQRDP.R
\name{plotSEIQRDP}
\alias{plotSEIQRDP}
\title{Plot active, recovered and deaths}
\usage{
plotSEIQRDP(
  object,
  active,
  recovered,
  deaths,
  title = NULL,
  params,
  checkRates = FALSE,
  ...
)
}
\arguments{
\item{object}{a SEIQRDP result}

\item{active}{time histories of the quarantined/active cases}

\item{recovered}{time histories of the recovered cases}

\item{deaths}{Time histories of the deceased cases}

\item{title}{country name}

\item{params}{fitted parameters by fit_SEIQRDP function}

\item{checkRates}{if \code{TRUE} compares the fitted and calcualted death and recovered ratios through plots}

\item{...}{other plot options}
}
\description{
This function computes the matrix A that is found in: dY/dt = A*Y + F
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
