% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQRDP_for_fitting.R
\name{SEIQRDP_for_fitting}
\alias{SEIQRDP_for_fitting}
\title{Fitted Results for SEIQRDP}
\usage{
SEIQRDP_for_fitting(par, t, t0, Npop, E0, I0, Q, R, D, dt)
}
\arguments{
\item{par}{initial guess parameters}

\item{t}{historical time vector}

\item{t0}{target time vector}

\item{Npop}{total population of the country}

\item{E0}{initial number of exposed cases}

\item{I0}{initial number of infectious cases}

\item{Q}{actual number of quarantined cases}

\item{R}{actual number of recovered cases}

\item{D}{actual number of dead cases}

\item{dt}{the time step. This oversamples time to ensure that the algorithm converges}
}
\value{
a data frame for fitted quarantined, recovered and deaths
}
\description{
Fitted Results for SEIQRDP
}
\references{
Peng, L., Yang, W., Zhang, D., Zhuge, C., Hong, L. 2020. “Epidemic analysis of COVID-19 in China by dynamical modeling”, arXiv preprint arXiv:2002.06563.

\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{fit_SEIQRDP}}  \code{\link{RK4}}
}
\author{
Selcuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
