% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRates.R
\name{checkRates}
\alias{checkRates}
\title{Check Rates}
\usage{
checkRates(time, Q, R, D, kappaFun, lambdaFun, kappa, lambda, dt = 1)
}
\arguments{
\item{time}{Time vector}

\item{Q}{Time histories of the quarantined/active cases}

\item{R}{Time histories of the recovered cases}

\item{D}{Time histories of the deceased cases}

\item{kappaFun}{Anonymous function approximating the death rate}

\item{lambdaFun}{Anonymous function approximating the recovery rate}

\item{kappa}{mortality rate}

\item{lambda}{cure rate}

\item{dt}{Time step, default is 1}
}
\value{
Creates plots for death rate and recovery rate. The idea is to check whether the approximation of these ratios is appropriate.
}
\description{
This function compares the fitted and calcualted death and recovered ratios.
The idea is to check whether the approximation of these ratios is appropriate
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
