% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SEIQRDP.R
\name{fit_SEIQRDP}
\alias{fit_SEIQRDP}
\title{Fit SEIQRDP function}
\usage{
fit_SEIQRDP(
  Q,
  R,
  D,
  Npop,
  E0,
  I0,
  time,
  guess,
  ftol = 1e-06,
  ptol = 1e-06,
  gtol = 1e-06,
  epsfcn = 0.001,
  factor = 100,
  maxfev = 1000,
  maxiter = 100,
  nprint = 1,
  trace = TRUE,
  ...
)
}
\arguments{
\item{Q}{time histories of the quarantined/active cases}

\item{R}{time histories of the recovered cases}

\item{D}{time histories of the deceased cases}

\item{Npop}{total population of the sample}

\item{E0}{initial number of exposed cases}

\item{I0}{initial number of infectious cases}

\item{time}{time vector}

\item{guess}{initiail guess parameters}

\item{ftol}{nls.lm.control object. non-negative numeric. Default is \code{1e-6}}

\item{ptol}{nls.lm.control object. non-negative numeric. Default is \code{1e-6}}

\item{gtol}{nls.lm.control object. non-negative numeric. Default is \code{1e-6}}

\item{epsfcn}{nls.lm.control object. Default is \code{0.001}}

\item{factor}{nls.lm.control object. Default is \code{100}}

\item{maxfev}{nls.lm.control object. Default is \code{1000}}

\item{maxiter}{nls.lm.control object. Default is \code{100}}

\item{nprint}{nls.lm.control object. Default is \code{1}}

\item{trace}{set \code{TRUE} to trace iteration results}

\item{...}{further arguments}
}
\description{
Fit SEIQRDP function parameters used in the SEIQRDP function, used to model
the time-evolution of an epidemic outbreak.
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
