% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA.R
\name{getA}
\alias{getA}
\title{Compute the matrix A}
\usage{
getA(alpha, gamma, delta, lambda, kappa)
}
\arguments{
\item{alpha}{protection rate}

\item{gamma}{inverse of the average latent time}

\item{delta}{rate of people entering in quarantine}

\item{lambda}{cure rate}

\item{kappa}{mortality rate}
}
\description{
This function computes the matrix A that is found in: dY/dt = A*Y + F
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
