% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataCOVID.R
\name{getDataCOVID}
\alias{getDataCOVID}
\title{Get COVID-19 Data}
\usage{
getDataCOVID(start = NULL, finish = NULL, country = NULL)
}
\arguments{
\item{start}{protection rate}

\item{finish}{inverse of the average latent time}

\item{country}{rate of people entering in quarantine}
}
\description{
The function collects the updated  COVID-19 data from the
John Hopkins University
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
