% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLambdaFun.R
\name{getLambdaFun}
\alias{getLambdaFun}
\title{Estimate Death Rate}
\usage{
getLambdaFun(
  tTarget,
  Q,
  R,
  guess,
  ftol,
  ptol,
  gtol,
  epsfcn,
  factor,
  maxfev,
  maxiter,
  nprint,
  trace
)
}
\arguments{
\item{tTarget}{time vector}

\item{Q}{target time-histories of the quarantined cases}

\item{R}{target time-histories of the recovered cases}

\item{guess}{Initiail guess parameters for kappa}

\item{ftol}{nls.lm.control object. non-negative numeric. Default is \code{1e-6}}

\item{ptol}{nls.lm.control object. non-negative numeric. Default is \code{1e-6}}

\item{gtol}{nls.lm.control object. non-negative numeric. Default is \code{1e-6}}

\item{epsfcn}{nls.lm.control object. Default is \code{0.001}}

\item{factor}{nls.lm.control object. Default is \code{100}}

\item{maxfev}{nls.lm.control object. Default is \code{1000}}

\item{maxiter}{nls.lm.control object. Default is \code{100}}

\item{nprint}{nls.lm.control object. Default is \code{1}}

\item{trace}{Set \code{TRUE} to trace iteration results}
}
\description{
This function provides a first estimate of the  death rate, to faciliate
convergence of the main algorithm.
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
