% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RK4.R
\name{RK4}
\alias{RK4}
\title{Runge-Kutta of order 4}
\usage{
RK4(Y, A, K, dt)
}
\arguments{
\item{Y}{time vector}

\item{A}{target time-histories of the quarantined cases}

\item{K}{target time-histories of the recovered cases}

\item{dt}{Initiail guess parameters for kappa}
}
\description{
Runge-Kutta of order 4
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{SEIQRDP}} \code{\link{fit_SEIQRDP}}
}
\author{
Selçuk KORKMAZ, \email{selcukorkmaz@gmail.com}
}
