% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQRDP.R
\name{SEIQRDP}
\alias{SEIQRDP}
\title{Simulate generalized SEIR model}
\usage{
SEIQRDP(
  alpha,
  beta,
  gamma,
  delta,
  lambda0,
  kappa0,
  Npop,
  E0,
  I0,
  Q0,
  R0,
  D0,
  lambdaFun,
  kappaFun,
  tstart,
  tfinish,
  dt = 1,
  f = 0
)
}
\arguments{
\item{alpha}{fitted protection rate}

\item{beta}{fitted  infection rate}

\item{gamma}{fitted  Inverse of the average latent time}

\item{delta}{fitted  rate at which people enter in quarantine}

\item{lambda0}{fitted  cure rate}

\item{kappa0}{fitted  mortality rate}

\item{Npop}{Total population of the sample}

\item{E0}{Initial number of exposed cases}

\item{I0}{Initial number of infectious cases}

\item{Q0}{Initial number of quarantined cases}

\item{R0}{Initial number of recovered cases}

\item{D0}{Initial number of dead cases}

\item{lambdaFun}{anonymous function giving the time-dependant recovery rate}

\item{kappaFun}{anonymous function giving the time-dependant death rate}

\item{tstart}{start date}

\item{tfinish}{finish date}

\item{dt}{time step}

\item{f}{future predictions}
}
\description{
This function simulates the time-histories of an epidemic outbreak using a generalized SEIR model
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
