% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQRDP_for_fitting.R
\name{SEIQRDP_for_fitting}
\alias{SEIQRDP_for_fitting}
\title{Nested Function}
\usage{
SEIQRDP_for_fitting(par, t, t0)
}
\arguments{
\item{par}{fitted protection rate}

\item{t}{fitted  infection rate}

\item{t0}{fitted  Inverse of the average latent time}
}
\description{
Nested Function
}
\references{
\url{https://www.mathworks.com/matlabcentral/fileexchange/74545-generalized-seir-epidemic-model-fitting-and-computation}
}
\seealso{
\code{\link{fit_SEIQRDP}}
}
\author{
Selçuk Korkmaz, \email{selcukorkmaz@gmail.com}
}
